Remove unused ant id, decrease size of Ant struct.

+++ ant.c
@@ -2,7 +2,7 @@
 #include <stdio.h> // for NULL
 #include <assert.h>
 
-inline guint ant_get_owner (Ant *ant)
+inline guint8 ant_get_owner (Ant *ant)
 {
 	assert (ant != NULL);
 	return ant->owner;
@@ -14,7 +14,7 @@ inline gchar ant_get_direction (Ant *ant)
 	return ant->direction;
 }
 
-inline void ant_set_owner (Ant *ant, guint player_id)
+inline void ant_set_owner (Ant *ant, guint8 player_id)
 {
 	assert (ant != NULL);
 	ant->owner = player_id;
